+++ mod-pubsub-server/src/main/java/org/folio/services/impl/KafkaTopicServiceImpl.java
@@ -33,14 +33,14 @@ public class KafkaTopicServiceImpl implements KafkaTopicService {
   public Future<Boolean> createTopics(List<String> eventTypes, String tenantId) {
     Promise<Boolean> promise = Promise.promise();
     List<NewTopic> topics = eventTypes.stream()
-      .map(eventType -> new NewTopic(new PubSubConfig(tenantId, eventType).getTopicName(), kafkaConfig.getNumberOfPartitions(), (short) kafkaConfig.getReplicationFactor()))
+      .map(eventType -> new NewTopic(new PubSubConfig(kafkaConfig.getEnvId(), tenantId, eventType).getTopicName(), kafkaConfig.getNumberOfPartitions(), (short) kafkaConfig.getReplicationFactor()))
       .collect(Collectors.toList());
     kafkaAdminClient.createTopics(topics, ar -> {
       if (ar.succeeded()) {
         LOGGER.info("Created topics: [{}]", StringUtils.join(eventTypes, ","));
         promise.complete(true);
       } else {
-        LOGGER.error("Some of the topics [{}] were not created. Cause: {}", StringUtils.join(eventTypes, ","), ar.cause().getMessage());
+        LOGGER.info("Some of the topics [{}] were not created. Cause: {}", StringUtils.join(eventTypes, ","), ar.cause().getMessage());
         promise.complete(false);
       }
     });