+++ mod-pubsub-server/src/main/java/org/folio/services/impl/SecurityManagerImpl.java
@@ -1,11 +1,7 @@
 package org.folio.services.impl;
 
-import static io.vertx.core.http.HttpMethod.PUT;
-import static io.vertx.core.json.Json.encode;
 import static java.lang.String.format;
 import static org.apache.commons.lang3.StringUtils.EMPTY;
-import static org.apache.commons.lang3.StringUtils.isNotBlank;
-import static org.folio.HttpStatus.HTTP_NO_CONTENT;
 import static org.folio.rest.util.OkapiConnectionParams.OKAPI_TOKEN_HEADER;
 import static org.folio.rest.util.RestUtil.doRequest;
 
@@ -20,9 +16,9 @@ import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.folio.HttpStatus;
 import org.folio.dao.PubSubUserDao;
-import org.folio.representation.User;
 import org.folio.rest.util.OkapiConnectionParams;
 import org.folio.services.SecurityManager;
+import org.folio.services.cache.Cache;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -50,14 +46,15 @@ public class SecurityManagerImpl implements SecurityManager {
   private static final String PERMISSIONS_FILE_PATH = "permissions/pubsub-user-permissions.csv";
   private static final String PUB_SUB_USERNAME = "pub-sub";
   private static final String USER_LAST_NAME = "System";
-  private static final String TOKEN_KEY_FORMAT = "%s_JWTToken";
 
   private final PubSubUserDao pubSubUserDao;
   private final Vertx vertx;
+  private final Cache cache;
 
-  public SecurityManagerImpl(@Autowired PubSubUserDao pubSubUserDao, @Autowired Vertx vertx) {
+  public SecurityManagerImpl(@Autowired PubSubUserDao pubSubUserDao, @Autowired Vertx vertx, @Autowired Cache cache) {
     this.pubSubUserDao = pubSubUserDao;
     this.vertx = vertx;
+    this.cache = cache;
   }
 
   @Override
@@ -69,7 +66,7 @@ public class SecurityManagerImpl implements SecurityManager {
       .compose(response -> {
         if (response.statusCode() == HttpStatus.HTTP_CREATED.toInt()) {
           LOGGER.info("Logged in pub-sub user");
-          putTokenToVertxContext(response.getHeader(OKAPI_TOKEN_HEADER), params);
+          cache.addToken(params.getTenantId(), response.getHeader(OKAPI_TOKEN_HEADER));
           return Future.succeededFuture(true);
         }
         LOGGER.error("pub-sub user was not logged in, received status {}", response.statusCode());
@@ -79,10 +76,10 @@ public class SecurityManagerImpl implements SecurityManager {
 
   @Override
   public Future<String> getJWTToken(OkapiConnectionParams params) {
-    String token = vertx.getOrCreateContext().get(format(TOKEN_KEY_FORMAT, params.getTenantId()));
+    String token = cache.getToken(params.getTenantId());
     if (StringUtils.isEmpty(token)) {
       return loginPubSubUser(params)
-        .map(v -> vertx.getOrCreateContext().<String>get(format(TOKEN_KEY_FORMAT, params.getTenantId())));
+        .map(v -> cache.getToken(params.getTenantId()));
     }
     return Future.succeededFuture(token);
   }
@@ -221,7 +218,7 @@ public class SecurityManagerImpl implements SecurityManager {
     String permUrl = PERMISSIONS_URL + "/" + userId + "/permissions?indexField=userId";
     permissions.forEach(permission -> {
       JsonObject requestBody = new JsonObject()
-        .put("permissionName",permission);
+        .put("permissionName", permission);
       futures.add(doRequest(requestBody.encode(), permUrl, HttpMethod.POST, params)
         .compose(response -> {
           Promise<Boolean> promise = Promise.promise();
@@ -248,10 +245,6 @@ public class SecurityManagerImpl implements SecurityManager {
       LOGGER.error("Error reading permissions from {}", e, path);
     }
     return permissions;
-  }
-
-  private void putTokenToVertxContext(String token, OkapiConnectionParams params) {
-    vertx.getOrCreateContext().put(format(TOKEN_KEY_FORMAT, params.getTenantId()), token);
   }
 
   private User createUserObject() {